+++ docs/source/releasenotes.rst
@@ -6,6 +6,32 @@ Release notes
 +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 
+**RPA.Assistant** (``rpaframework-assistant`` **2.1.0**):
+
+    - Features:
+
+        - Add ``Set Title`` keyword. Can be used to set title of assistant when it is running.
+
+        - Add ``Open Row`` and ``Close Row`` keywords. Used to layout elements into rows.
+        - Add ``Open Column`` and ``Close Column`` keywords. Used to layout elements into columns.
+        - Add ``Open Navbar`` and ``Close Navbar`` keywords. Can be used to create an always visible top bar for a dialog.
+        
+        - Add ``Open Stack`` and ``Close Stack`` keywords. Can be used to position elements manually.
+        - Add ``Open Container`` and ``Close Container`` keywords. Useful for styling or placing single elements.
+
+        - Add ``Add Flet Icon`` that enables adding any icons from flets large gallery of icons
+
+            - Compared to `Add Icon` it is more difficult to use but supports a much larger amount of icons.
+
+    - Bugfixes:
+
+        - Fix regression, Assistant flet app not closing properly when run from CLI and
+          when close button is pressed.
+
+        - The library no longer silently misses adding to the result when an input
+          ``name`` duplicate name is used, instead raises a ValueError.
+
+
 `Released <https://pypi.org/project/rpaframework/#history>`_
 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++